Removal of path.join for cross-platform compatibility
Currently there is a bug in Windows where all of the expanded aliases will have paths with backslashes in it.  This is due to the use of path.join to create the alias, which uses the current platform's path separator.

Since Glob is returning paths with forward slash - even on Windows - the solution is to put `pattern.expose + '/'` in front of the file when pattern.expose is present, or just using the file unhindered as the alias if pattern.expose is absent.