Removal of path.resolve to mask absolute paths
Currently when you create aliases, the aliases point to the full file path (ie: `c:\\dev\\workspace\\index.js`).  However, this can result in the full file path ending up in the merged file in browserify in some circumstances (rendr has this problem).

This simply changes the absolute path to a relative path based on the cwd option.